move more core modules to their path
