+++ template/.gitignore.j2
@@ -36,7 +36,6 @@ assets/statics/js/bundle.js.map
 
 # App Specific
 ############################################################
-src/{{ project_slug }}/_version.py
 
 # Custom ignored patterns
 # (Add any app specific ignored patterns here)
