+++ template/pyproject.toml.j2
@@ -61,7 +61,7 @@ fallback-version = "0.0.0"
 source = "uv-dynamic-versioning"
 
 [tool.hatch.build.hooks.version]
-path = "src/{{ project_slug }}/_version.py"
+path = "src/{{ project_slug }}/core/_version.py"
 template = '''
 version = "{version}"
 '''
