+++ template/src/{{ project_slug }}/core/__init__.py
@@ -1,11 +1,14 @@
-from . import paths
+from . import paths, templates
 from .config import project_settings, settings
+from .context import Context
 from .logging import logger
 
 
 __all__ = [
+    "Context",
     "logger",
     "paths",
-    "settings",
     "project_settings",
+    "settings",
+    "templates",
 ]
