+++ template/src/{{ project_slug }}/core/_version.py
@@ -1,0 +1,1 @@
+version = "0.0.0-default"
