+++ template/src/{{ project_slug }}/core/config.py
@@ -17,8 +17,8 @@ from pydantic import (
 from pydantic_extra_types.language_code import LanguageAlpha2
 from pydantic_settings import BaseSettings, SettingsConfigDict
 
-from .._version import version
 from . import paths
+from ._version import version
 
 
 current_year: int = datetime.now().year
