+++ template/src/{{ project_slug }}/core/time.py
