+++ template/src/{{ project_slug }}/frontend/context.py
@@ -1,4 +1,4 @@
-from ..context import Context
+from ..core import Context
 
 
 class AppContext(Context):
