+++ template/src/{{ project_slug }}/frontend/templates.py
@@ -5,9 +5,9 @@ from fastapi.templating import Jinja2Templates
 from starlette.responses import HTMLResponse
 from starlette_babel.contrib.jinja import configure_jinja_env
 
-from ..context import Context
+from ..core import Context
 from ..core.paths import frontend_templates
-from ..templates import render_static_template
+from ..core.templates import render_static_template
 from .hotreload import hotreload
